requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
allantools>=1.0.0
PyYAML>=6.0
psutil>=5.8.0

# Hardware dependencies (optional - will gracefully degrade if not available)
RPi.GPIO>=0.7.1
rplcd>=1.2.0

# Web automation dependencies
playwright>=1.40.0
requests>=2.28.0

# Development dependencies (optional)
pytest>=6.0.0
pytest-cov>=2.12.0
